requirements.txt
streamlit==1.29.0
streamlit-autorefresh==0.0.1
requests==2.31.0
pandas==2.1.4
altair==5.2.0
twilio==8.10.0
python-dotenv==1.0.0
pydantic==2.5.2
redis==5.0.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
celery==5.3.4
prometheus-client==0.19.0
structlog==23.2.0